Move mobile messages feature to messages/mobile
