+++ messages/mobile/src/androidMain/kotlin/com/boswelja/smartwatchextensions/messages/DI.kt
@@ -3,12 +3,14 @@ package com.boswelja.smartwatchextensions.messages
 import com.boswelja.smartwatchextensions.messages.ui.MessageHistoryViewModel
 import com.boswelja.smartwatchextensions.messages.ui.MessagesViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.core.context.loadKoinModules
 import org.koin.dsl.module
 
 /**
  * A Koin module for providing messages classes.
  */
-val messagesUiModule = module {
+val messagesModule = module {
+    loadKoinModules(messagesCommonModule)
     viewModel { MessagesViewModel(get(), get()) }
     viewModel { MessageHistoryViewModel(get()) }
 }
