+++ messages/mobile/src/androidMain/kotlin/com/boswelja/smartwatchextensions/messages/ui/MessagesViewModel.kt
@@ -2,9 +2,9 @@ package com.boswelja.smartwatchextensions.messages.ui
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
+import com.boswelja.smartwatchextensions.core.UpdateChecker
 import com.boswelja.smartwatchextensions.messages.DisplayMessage
 import com.boswelja.smartwatchextensions.messages.MessagesRepository
-import com.boswelja.smartwatchextensions.updatechecker.UpdateChecker
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.withContext
