+++ messages/mobile/src/androidMain/res/values/strings.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <!-- Messages strings -->
+    <string name="message_history_label">Message History</string>
+    <string name="messages_empty">No new messages</string>
+    <string name="messages_noti_channel_label">System Messages</string>
+    <string name="message_history_empty">Messages you\'ve dismissed will appear here.</string>
+    <string name="message_action_changelog">See Changes</string>
+    <string name="message_action_noti_settings">Notification Settings</string>
+    <string name="message_action_install_update">Install Update</string>
+    <string name="message_dismissed">Message Dismissed</string>
+    <string name="message_history_cleared">Message history cleared</string>
+    <string name="message_history_clear_all">Clear All</string>
+
+</resources>
